+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - JSON result will have `ok` boolean as a helper, it will be `true` if there is any drift or drift detection failed, `false` otherwise.
+- JSON result will have the workspace tags as a helper to grab information.
 
 ## [v0.2.0] - 2022-11-15